solana_packet

Struct Packet

#[repr(C)]
pub struct Packet { /* private fields */ }

Implementations

impl Packet

pub fn new(buffer: [u8; 1232], meta: Meta) -> Self

pub fn data<I>(&self, index: I) -> Option<&<I as SliceIndex<[u8]>>::Output>
where I: SliceIndex<[u8]>,

Returns an immutable reference to the underlying buffer up to packet.meta.size. The rest of the buffer is not valid to read from. packet.data(..) returns packet.buffer.get(..packet.meta.size). Returns None if the index is invalid or if the packet is already marked as discard.

pub fn buffer_mut(&mut self) -> &mut [u8]

Returns a mutable reference to the entirety of the underlying buffer to write into. The caller is responsible for updating Packet.meta.size after writing to the buffer.

pub fn meta(&self) -> &Meta

pub fn meta_mut(&mut self) -> &mut Meta

pub fn from_data<T: Encode>(dest: Option<&SocketAddr>, data: T) -> Result<Self>

pub fn populate_packet<T: Encode>( &mut self, dest: Option<&SocketAddr>, data: &T, ) -> Result<()>

pub fn deserialize_slice<T, I>(&self, index: I) -> Result<T>
where T: DeserializeOwned, I: SliceIndex<[u8], Output = [u8]>,